Removing resource_action_mixin.rb since it's only used by resource.rb
